working on device read/write signals
We will need to use the inverted outputs from the '175
flip flops latching the microcode signals in order to
ensure that the device-specific read/write signals are
high on reset.  This means using positive logic for
decoding the device ids.  NOR gates do the trick for
the GP registers, and the '238 decoder does the other
devices (it produces active high outputs, rather than
active low like the '138 decoder.)