eliminated pcClkEn signal
There's no need to selectively enable/disable
the count function in the PC because the PC
clock signal (pcClk) is generated explicitly
by the microcode, and clock pulses are sent
only when we actually want to increment the PC
value.

This allowed a simplicification of the carry
logic, and eliminated the need for the 74HC30
in the last stage of the carry logic.