fleshing out full set of microcode signals
This won't be the final version: the signals for
reading from a device will be different.  There is no
need to have separate signals for reading the high
and low registers for register pairs (e.g., PCHi
and PCLo) because the ALU can select which one should
be sent to the ALU data bus.  So, only 3 bits are needed
to specify which device is being read, as long as
the appropriate ALU operation is asserted to select
the data from the left or right ALU bus.