construct microcode words from MSB to LSB
So, the first signal is the most significant.

This makes the ordering of the bitstrings consistent
with the overall instruction word.  This also makes
sure that the individual bits in each byte match the
order of the ROM outputs.

This does mean that the first ROM is actually
the high-order byte of the microcode word.