more redesign of microcode
Reading and writing devices and ALU control are
mostly in place.

Eliminated the dedicated signals to assert device pairs
to the left/right ALU input busses.  These aren't needed
any more because all reading from devices is done using
pairs (with the ALU to select which one is copied to the
ALU data bus.)