rotating the ucode words is no longer necessary
The initial negative clock pulse clocks the 0 value
from the '590 counter into the output register,
from which it then selects microcode word 0 from the
ROMs, and then the data is clocked into the '175
registers on the first positive clock pulse.  The
rotation was necessary when we were using '393 counters
where the first negative clock pulse set the counter
to 1 to select the first data value from the ROMs.