have AR regs output to the low/high addr busses
For store instructions, a GP register pair is used
to specify the address, AND a GP register specifies
the data.  This means we can't drive the address
bus directly from a GP register pair, since that
wouldn't allow us to output the correct data.  So,
the contents of the GP register pair specifying the
address must be transferred into the AR registers,
which can then drive the address, allowing the
correct GP register data to be output via the ALU.

Will need to add a microcode signal for driving the
address bus from the AR registers.