started template for store to memory
It's complicated because it involves a transfer
from a GP register pair to AR, but I think it should
be on the right track.

It occurs to me that we will need external memory
read/write signals.  Shouldn't be a problem, since
we have unused microcode bits and unused '175 flip
flops.

Note about previous commit: we actually don't need a
new microcode signal for asserting AR onto the address
busses, since the existing rdAR signal can do this.
(Note that we lose the ability to read the AR regs
onto the left/right ALU input busses, but that should
be ok.)