generate external r/w signals
Found a way to do this without increasing the chip
count: use NAND gates to produce some of the conditional
end signals (rather than OR gates), and use AND gates
to generate the -ucodeRST signal and the new
extW/extR signals.  These are active high, and
are generated as inverted, ensuring that the
final -uExtW and -uExtR signals are not asserted
on reset.