add explicit signal to latch condition codes
Was able to use microcode bits more efficiently (to make
room for the new signal) by putting the combinational
logic to generate the external read/write signals after
the '175 microcode signal registers.  This should be fine
as long as memDir is set appropriately before the extRW
signal is asserted.  (No glitches should be possible.)