account for opfetch for -uExtR and -uRwMem
-uExtR should be asserted whenever -opReadMem is asserted,
and -uRwMem should be asserted whenever -opfetch is asserted.

Did this without additional chips, at the cost of requiring
74HCT10 rather than 74HCT20 in the conditional branch logic.
(Will have to buy these, but I've seen them on ebay.)

Also renumbered some of the chips in the ucode2 module
so that the numbering reflects the actual number of chips
(16).

Will need to carefully double check this to make sure it's
correct.